Renamed and moved implementing class
